requirements.txt
asgiref==3.8.1
pytz==2024.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=4.3.0,<5.0.0
djoser>=2.1.0,<3.0
